update alerts to sweet alerts in terms.

+++ scripts/term.js
@@ -62,7 +62,7 @@ function addTerm() {
   })
   .then((response) => response.text())
   .then(responseText => {
-    alert(responseText);
+    Swal.fire("Success", `${responseText}`, "success");
     displayTerms(); // updates the table
   }).catch (error => {
     alert('console error.');
@@ -107,7 +107,7 @@ function editTerm(button) {
     })
     .then((response) => response.text())
     .then((responseText) => {
-      alert(responseText);
+      Swal.fire("Success", `${responseText}`, "success");
       displayTerms();
     });
 
@@ -116,6 +116,28 @@ function editTerm(button) {
     // Edit mode
     button.textContent = "Save";
     cells[0].focus();
+
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      icon: "info",
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      customClass: {
+        popup: 'colored-toast'
+      },
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      }
+    });
+
+    Toast.fire({
+      icon: 'info',
+      title: 'Edit Mode Activated'
+    });
+
     updatedRow.length = 0; // clear the array.
   }
 }
@@ -125,19 +147,28 @@ function deleteTerm(button) {
   const row = button.closest('tr');
   const cells = row.querySelectorAll('td');
   const termId = cells[0].innerHTML;
-  
-  fetch(termEndpoint, {
-    method: 'DELETE',
-    headers: {
-      "Content-type": "application/x-www-form-urlencoded",
-    },
-    body: `term_id=${termId}`,
-  })
-  .then((response) => response.text())
-  .then((responseText) => {
-    alert(responseText);
-    displayTerms();
-  })
+
+  Swal.fire({
+    title: "Do you want to delete this data?",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonText: "Delete",
+  }).then((result) => {
+    if (result.isConfirmed) {
+        fetch(termEndpoint, {
+          method: 'DELETE',
+          headers: {
+            "Content-type": "application/x-www-form-urlencoded",
+          },
+          body: `term_id=${termId}`,
+        })
+        .then((response) => response.text())
+        .then(() => {
+          Swal.fire("Deleted", "", "success");
+          displayTerms();
+      })
+    }
+  });
 }
 
 function searchTerm() {
